minBal to use double instead of BigDecimal when querying

+++ src/main/java/com/badals/shop/repository/search/OrderSearchRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface OrderSearchRepository extends ElasticsearchRepository<OrderDTO, Long> {
    //AddProductDTO findBySlug(String slug);
    List<OrderDTO> findAllByCustomerContainsAndTenantIdOrderById(String search, String tenantId, Pageable p);
-   List<OrderDTO> findAllByOrderStateInAndBalanceGreaterThanEqualAndTenantIdOrderByInvoiceDateAsc(List<OrderState> orderStates, BigDecimal balance, String tenantId, Pageable p);
-   List<OrderDTO> findAllByOrderStateInAndBalanceGreaterThanEqualAndTenantIdOrderByInvoiceDateDesc(List<OrderState> orderStates, BigDecimal balance, String tenantId, Pageable p);
+   List<OrderDTO> findAllByOrderStateInAndBalanceGreaterThanEqualAndTenantIdOrderByInvoiceDateAsc(List<OrderState> orderStates, Double balance, String tenantId, Pageable p);
+   List<OrderDTO> findAllByOrderStateInAndBalanceGreaterThanEqualAndTenantIdOrderByInvoiceDateDesc(List<OrderState> orderStates, Double balance, String tenantId, Pageable p);
 
 }
